ref(internal): move GetHTTPSpanName to internal package
