--- http/route_name_below_go1.23.go
@@ -1,9 +1,0 @@
-//go:build !go1.23
-
-package sentryhttp
-
-import "net/http"
-
-func getHTTPSpanName(r *http.Request) string {
-	return r.Method + " " + r.URL.Path
-}
