+++ internal/traceutils/route_name_below_go1.23.go
@@ -1,0 +1,10 @@
+//go:build !go1.23
+
+package traceutils
+
+import "net/http"
+
+// GetHTTPSpanName grab needed fields from *http.Request to generate a span name for `http.server` span op.
+func GetHTTPSpanName(r *http.Request) string {
+	return r.Method + " " + r.URL.Path
+}
